Started on pan sensor (Minor sync)
A sensor should now be on top of the pans other fixtures in the same
body, this is for collision with batter and pancakes to either cook or
create them.
BodyData is now used as the userdata object in the bodies of entities -
this means that you can get the entity from the body (great for checking
entity properties when collision detecting).